+++ src/app/Util/Enums.ts
@@ -1,0 +1,7 @@
+export enum Screens {
+  Home = '/',
+  Scheduled = '/scheduled',
+  RemindoroInfo = '/remindoro-info/:id',
+  Settings = '/settings',
+  Help = '/help',
+}