+++ src/app/Util/index.test.ts
@@ -2,5 +2,7 @@ import { getRemindoroUrl } from './'
 
 describe('Remindoro url is constructed correctly', () => {
   const id = 'porumai'
-  expect(getRemindoroUrl(id)).toEqual(`/remindoro-info/${id}`)
+  test('Url construction is fine', () => {
+    expect(getRemindoroUrl(id)).toEqual(`/remindoro-info/${id}`)
+  })
 })